Implement useCelculateHeight to get height
